refactor: change naming; extract ListInput element
