+++ todo-app/src/TodoList.js
@@ -1,38 +1,25 @@
 import React, { useState } from 'react';
-import TodoItems from './TodoItems';
-
-// const todos = []
-
+import TodoItems from './TodoItem';
+import ListInput from './ListInput';
+import ListBody from './ListBody';
 
 const TodoList = () => {
+    //everytime it refreshes, todos is set to []
 
     const [todos, setTodos] = useState([])
-
-    const addNewTodo = (e) => {
-        //When user presses "Enter" key, a new todo is created 
-        if (e.key === "Enter") {
-            const newTodo = e.target.value
-            //User cannot create todo with empty string nor dulipcated todo
-            if (newTodo.length !== 0 && !isDuplicatedTodo(newTodo)) {
-                setTodos(todos.concat(newTodo))
-
-                //empty input area
-                e.target.value = ""
-            }
-        }
-    }
-    const isDuplicatedTodo = (todo) => {
-        return todos.includes(todo)
-    }
 
     return (
         < div className="row" >
             <div className="col-3">
                 <ListHeader />
-                <div className="mb-3">
-                    <input type="text" className="form-control" id="addTodo" placeholder="today" onKeyPress={(e) => { addNewTodo(e) }} />
-                </div>
-                <ListContent todos={todos} />
+                <ListInput
+                    todos={todos}
+                    setTodos={setTodos}
+                />
+                < ListBody
+                    todos={todos}
+                    setTodos={setTodos}
+                />
             </div>
             <div className="col-3">Tomorrow</div>
             <div className="col-3">This Week</div>
@@ -43,29 +30,9 @@ const TodoList = () => {
 }
 
 
-const ListContent = (props) => {
-    const todos = props.todos
-
-    if (todos.length > 0) {
-        return (<TodoItems todos={todos} />)
-    } else {
-        return (<EmptyTodo />)
-    }
-}
 const ListHeader = () => {
     return <h5>Today</h5>
 }
-
-
-
-const EmptyTodo = () => {
-    return (
-        <div>
-            <span> You do not have any todos yet.</span >
-        </div >
-    )
-}
-
